feat(GUI): Added a simple GUI for selecting data
The GUI has a bottom and label. When pressing the bottom it triggers the file explorer with the mandatory filters mentioned in DoD, showing exclusively:
.csv
.xlsx, .xls
.sqlite, .db
and an extra option to select only one of those extension types. It replace the label previously created with the file path once selected. The data_module was slightly modified to succesfully connect it to the GUI

DoD: "Definición de Terminado", GUI: "Graphical user interface", Label: "Text shown"