perf(GUI): improved a lot time efficiency
It reduces drastically the time for showing the data on screen, from aproximately 14s -> 0.1s. This was achieved changing the implementation of the table, instead of using Qtablewidget which creates an object per cel(with 24 thousand lines of code and 10 columns is almost 240000 objects every time a dataframe shows on screen), we used Qtableview.
But this code will need refactoring in the future for sure.